MOVE; consumer.service, consumer.controller test files
to test folder
